concordia_moderator/pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"


[project]
name = "concordia-moderator"
version = "0.2.0"
description = "Concordia-based moderator system with real LLMs & embeddings"
authors = [{name = "Your Team"}]
requires-python = ">=3.9"
dependencies = [
"pyyaml>=6.0",
"numpy>=1.24",
"scikit-learn>=1.3",
]


[project.optional-dependencies]
openai = ["openai>=1.30.0"]
sentencetfm = ["sentence-transformers>=3.0.0"]
faiss = ["faiss-cpu>=1.7.4"]
all = ["openai>=1.30.0", "sentence-transformers>=3.0.0", "faiss-cpu>=1.7.4"]


[tool.pytest.ini_options]
addopts = "-q"